W04 Team Activity: Checkout. Updates
