+++ src/css/style.css
@@ -300,38 +300,103 @@ h2 {
   text-align: center;
 } */
 
-h3 {
-  font-size: 24px;
-  margin: 0 0 15px 0;
-  border-bottom: 1px solid #777;
-  padding: 0;
-  text-align: center;
-}
-
-label {
-  display: block;
-  margin: 5px;
-}
-
-input {
-  padding: .25rem;
-  width: 100px;
+.checkout-container {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  max-width: 900px;
+  width: 100%;
+  margin: 0 auto;
+  gap: 20px;
 }
 
 .checkout-form {
+  display: flex;
+  justify-content: center;
+  background-color: #f5f5f5;
+  font-family: Arial, sans-serif;
+  margin: 0;
+  padding: 20px;
+}
+
+.checkout-form {
+  display: flex;
+  flex-direction: column;
+  max-width: 800px;
+  width: 100%;
+  background-color: #fff;
+  padding: 20px;
+  border-radius: 5px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+}
+
+fieldset {
+  border: 1px solid #ddd;
+  padding: 20px;
+  margin-bottom: 20px;
+  border-radius: 5px;
+}
+
+legend {
+  padding: 0 10px;
+  font-weight: bold;
+}
+
+.checkoutName,
+.checkoutAddress,
+.paymentMethod {
   display: grid;
   grid-template-columns: 1fr 1fr;
   grid-gap: .25rem;
 }
 
-.checkout-form label {
-  text-align: right;
+label {
+  display: block;
+  margin-bottom: 10px;
+  font-weight: bold;
+}
+
+input {
+  width: calc(100% - 20px);
+  padding: 10px;
+  margin-bottom: 20px;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+}
+
+.checkout-summary {
+  /* position: absolute; */
+  right: 20px;
+  top: 20px;
+  max-width: 300px;
+  width: 100%;
+}
+
+.checkout-summary ul {
+  list-style-type: none;
+  padding: 0;
+}
+
+.checkout-summary li {
+  margin-bottom: 10px;
+}
+
+.checkout-summary p {
+  display: inline;
+  margin-left: 10px;
 }
 
 button {
-  padding: .5rem;
-  margin: 10px auto;
+  padding: 10px 20px;
+  background-color: #007bff;
+  color: #fff;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 16px;
+}
 
+button:hover {
+  background-color: #0056b3;
 }
 
 @media screen and (min-width: 500px) {
