+++ src/js/checkout.js
@@ -1,3 +1,6 @@
 import { loadHeaderFooter } from "./utils.mjs";
+//import { CheckoutProcess } from "./CheckoutProcess.mjs";
 
 loadHeaderFooter();
+
+
